requirements.txt
gcloud==0.18.3
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-bigquery==3.9.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
requests==2.25.1
